+++ src/settings/ActionProfiles/ViewActionProfile.js
@@ -230,6 +230,7 @@ export class ViewActionProfile extends Component {
                 link={tagsEntityLink}
                 getEntity={getEntity}
                 getEntityTags={getEntityTags}
+                entityTagsPath="profile.tags"
               />
             </div>
           )}