+++ src/utils/getDropdownOptions.js
@@ -1,18 +1,7 @@
 import { get } from 'lodash';
 
-import ordersTranslations from '@folio/orders/translations/ui-orders/en'; // eslint-disable-line import/no-unresolved
-import inventoryTranslations from '@folio/inventory/translations/ui-inventory/en'; // eslint-disable-line import/no-unresolved
-import invoiceTranslations from '@folio/invoice/translations/ui-invoice/en'; // eslint-disable-line import/no-unresolved
 import { fieldsConfig } from './fields-config';
 import { fieldCategoriesConfig } from './field-categories-config';
-
-const translations = {
-  ...ordersTranslations,
-  ...inventoryTranslations,
-  ...invoiceTranslations,
-};
-
-export const getLabel = key => translations[key];
 
 export const matchFields = (resources, recordType) => {
   return fieldsConfig.filter(field => field.recordType
@@ -22,13 +11,16 @@ export const matchFields = (resources, recordType) => {
 
 export const getCategoryId = field => fieldCategoriesConfig.find(category => category.id === field.categoryId);
 
-export const getDropdownOptions = records => {
+export const getDropdownOptions = (records, intl) => {
   return records.map(record => {
     const category = getCategoryId(record);
 
+    const categoryValue = intl.formatMessage({ id: category.label });
+    const labelValue = intl.formatMessage({ id: record.label });
+
     return {
       value: record.value,
-      label: `${getLabel(category.label)}: ${getLabel(record.label)}`,
+      label: `${categoryValue}: ${labelValue}`,
     };
   });
 };