+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Handle import of stripes-acq-components to modules and platform (UISACQCOMP-3)
 * Add validation rules for Move action for the MARC modifications table fields (UIDATIMP-492)
 * Add validation rules for Edit action for the MARC modifications table fields (UIDATIMP-489)
+* Get rid of imported translations for math profiles (UIDATIMP-570)
 
 ### Bugs fixed:
 * Fix deletion repeatable fields in Field mapping profile (UIDATIMP-482)