+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 ### Bugs fixed:
 * Fix deletion repeatable fields in Field mapping profile (UIDATIMP-482)
+* Fix assigning and unassigning tags to data import profiles (UIDATIMP-499)
 
 ## [2.1.0](https://github.com/folio-org/ui-data-import/tree/v2.1.0) (2020-06-25)